+++ explainer/explainer_null_test.go
@@ -21,21 +21,3 @@ func TestExplainerNull(test *testing.T) {
 	err := testObject.Explain(ctx)
 	assert.Nil(test, err)
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleExplainerNull() {
-	// For more information, visit https://github.com/senzing-garage/explain/blob/main/explainer/explainer_error_test.go
-	ctx := context.TODO()
-	explainer := &explainer.ExplainerError{
-		ErrorId: "senzing-60010000",
-		TtyOnly: true,
-	}
-	err := explainer.Explain(ctx)
-	if err != nil {
-		panic(err)
-	}
-	//Output: For information on that error, visit https://hub.senzing.com/g2-sdk-go-base/errors#senzing-60010000
-}