+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2024-04-19
+
+### Changed in 0.2.1
+
+- Update dependencies
+  - github.com/senzing-garage/go-cmdhelping v0.2.1
+  - github.com/stretchr/testify v1.9.0
+
 ## [0.2.0] - 2024-01-02
 
 ### Changed in 0.2.0